% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/F_to_C.R
\name{F_to_C}
\alias{F_to_C}
\title{F_to_C}
\usage{
F_to_C(tempF)
}
\arguments{
\item{tempF, }{a temperature measurement in Fahrenheit, can be a single value or a vector. I chose tempF as it's a temperature value in F}
}
\value{
a numeric value or vector in Celsius
}
\description{
This function converts a temperature measurement from Fahrenheit to Celsius.
}
\examples{
F_to_C(32)
F_to_C(c(15, 89))
}
